Convert librawsock to single-header library with Chinese README
