--- include/config.h.in
@@ -1,23 +1,0 @@
-/**
- * @file config.h
- * @brief Configuration header for librawsock
- * @author Sphinxes0o0
- * @version @PROJECT_VERSION@
- * 
- * This file is automatically generated by CMake.
- */
-
-#ifndef LIBRAWSOCK_CONFIG_H
-#define LIBRAWSOCK_CONFIG_H
-
-/* Version information */
-#define RAWSOCK_VERSION_MAJOR @PROJECT_VERSION_MAJOR@
-#define RAWSOCK_VERSION_MINOR @PROJECT_VERSION_MINOR@
-#define RAWSOCK_VERSION_PATCH @PROJECT_VERSION_PATCH@
-#define RAWSOCK_VERSION_STRING "@PROJECT_VERSION@"
-
-/* Build configuration */
-#define RAWSOCK_BUILD_SHARED @BUILD_SHARED_LIBS@
-#define RAWSOCK_BUILD_STATIC @BUILD_STATIC_LIBS@
-
-#endif /* LIBRAWSOCK_CONFIG_H */
