--- librawsock.pc.in
@@ -1,10 +1,0 @@
-prefix=@CMAKE_INSTALL_PREFIX@
-exec_prefix=${prefix}
-includedir=${prefix}/include
-libdir=${exec_prefix}/lib
-
-Name: librawsock
-Description: Simple raw socket library
-Version: @PROJECT_VERSION@
-Cflags: -I${includedir}
-Libs: -L${libdir} -lrawsock
